zero bits for trailer
Use zero bits, not ones, for the trailer, to avoid lighting up
downstream pixels when the DotStar object is created with
fewer pixels than are physically present in the chian.

This is visible after the second update of a DotStar object that
is at least 17 pixels long, when there is a greater number of
physical pixels in the chain.

Signed-off-by: Taylor Yu <code@argon.blue>